more field->channel renaming
