+++ frontends/mit-open/src/pages/ChannelPage/EditChannelAppearanceForm.tsx
@@ -6,7 +6,7 @@ import * as Yup from "yup"
 
 import type { FieldChannel } from "api/v0"
 import { makeFieldViewPath } from "@/common/urls"
-import { useChannelPartialUpdate } from "api/hooks/fields"
+import { useChannelPartialUpdate } from "../../../../api/src/hooks/channels"
 
 type FormProps = {
   field: FieldChannel
